Handle packet size limit for acme-dns-01 correctly
There was an edge case here requiring: TCP transport (for full 16K
limit), the query was for an _acme-challenge hostname, a large
amount of static TXT data for the hostname in the zonefile, and a
a large amount of dynamic TXT data injected by acme-dns-01.  If
the net output added up to just under 16KB, but within the last 17
bytes of that, we'd do the Wrong Thing because this code looked at
the fixed MAX_RESPONSE rather than ctx->this_max_response.

While this fixes that immediate issue in the small view, it just
serves to highlight further that our hacky limited acme-dns-01
support should probably be more general along a couple of
dimensions and that the lookup code should probably be shifted
over to dnspacket.c as well.  We'll see if that makes the cut for
3.x or we just deal with it all Later.