libsodium integration
As of this commit, libsodium 1.x is a hard build requirement.

The primary immediate driver for this is to use it for the light
crypto bits involved in supporting EDNS Cookies properly.  Later,
if we pursue DNSSEC, we can likely build most of the crypto bits
of that with it as well.

And since we have libsodium available, I've gone ahead and killed
our hacky gdnsd_init_rand() and related functions in libgdnsd,
replacing them with initialization of our lightweight PRNG states
directly from libsodium's PRNG outputs, which kills a fair amount
of cruft in this area.

The sodium init function is called from a gdnsd function with
__attribute__((__constructor__)), which I believe is a fairly
standard thing on *nix/POSIX-y targets at this point, and gets it
initialized very early (before main()).