Fixup testsuite to use upstream Net::DNS again
Many, many versions ago, we forked an old copy of Net::DNS and
began including it in our testsuite directory, because upstream
Net::DNS broke interfaces we cared about and it was a pain to fix
it all.

The testsuite itself has evolved since then, and some upcoming
changes need more out of Net::DNS than our old frozen version can
offer.

With this change, the old frozen copy is deleted, and the
testsuite is updated to work with upstream Net::DNS again, which
is a build dependency if you want to run the testsuite.  This
requires the upstream version to be 1.03 or higher, and I've
tested that the testsuite works on every version from 1.03 to 1.19
(current as of this writing).  I tried moving the minimum
requirement even lower, but the bugs became too severe below 1.03
to be worth working around.

I wouldn't be shocked if future versions break us again, as it
took a fair amount of hackery just to paper over the various
interface breakages and bugs present in the 1.03 - 1.19 range.
Upstream probably doesn't care about stability in the kinds of
deep interfaces we're using; they're not focused on use-cases
where generating very specifically-encoded and/or malformed
packets for testsuites is a requirement.  I moved several such
tests away from Net::DNS using hardcoded binary packet strings in
hopes of avoiding future bugs, but we'll see how it goes!

In the long term, we should rewrite the testsuite using a
completely different harness and protocol implementation (which
probably means not Perl-based.  Maybe dnspython, or cook something
up using one of the C-based libraries).  But for now, this will
just have to do!