Improve hardening in autoconf
This gets rid of the _PIE variants of CFLAGS/LDFLAGS and uses the base
ones to set the hardening flags if support is detected, thus simplifying
a bit of autoconf junk.  This also gets fPIE turned on while building
archive members, which solves some mysteries around some clang PIE build
issues as well!