Add Werror testing to qa scripts
It's tricky to inject -Werror into CFLAGS optionally, as doing it from
outside of configure will screw up various configure checks, and we
don't want it on for all builds (because realistically, I can only
make warnings-clean code for common compilers I use on my machine, not
for all other build targets and compilers).

So, this adds a configure argument "--with-werror" that injects -Werror
into CFLAGS after all the other autoconf checks are done, and adds a QA
script which makes use of it against my local list of compilers.