Remove the recent auto-serial feature
I had added the ability to autogenerate zone serials from disk
mtimes in a standard format, if the actual serial number in the
zonefile was the magic number zero.

On further reflection, this is a mis-feature.  It *seems* like it
might be useful for some use-cases, but in practice eventually
everyone that tries to use it will migrate away from it once they
realize all the real-world operational problems it causes with a
real set of multiple production servers (think FS timestamps, and
reinstalled machines and adding new machines to the set, etc).

At the end of the day, consistent serials either need to be set
manually in upstream data (during git commits?) or need to be set
by the zonefile deployment tooling, to keep them consistent (and
yes, some registrars will complain if your serials aren't
consistent across authservers).

This is mostly a revert of db24c506 , but a few minor intervening
changes to related lines made it not a clean perfect revert.