Remove support for djbdns zonefiles
I've kept this alive for quite some time through quite a lot of
changes, mostly because it's the only example of a second zone
data backend alongside the canonical rfc1035 one, but now its time
has sadly come to an end.

The key rationale here is that there's really no way it's going to
cope with all the changes coming in future work (notably, DNSSEC
support), and I don't want to have to work around that or go
through another major-version cycle to delete it at that point.

The minor points that were already going against it: it's not been
really updated since ~2013 when it was first pulled in and marked
EXPERIMENTAL, it has only ever supported a minimal set of RR types
(just SOA, A, NS, MX, CNAME, TXT, but not PTR, AAAA, RFC 3597,
etc), and it has never had any decent testsuite coverage aside
from loading one trivial file and testing one A lookup against it.

Sorry, Timo!