Reduce test port usage
We actually only need 3 of the 5 ports we originally allocated to
each test number: HTTP is gone, and there was probably never a
good reason for splitting the IPv[46] source port numbers.

Combined with the renaming commit just before this one: the
testsuite was using 260 out of a range of 285 ports (due to holes
from deleted tests), and now it's fully using a set of 156 ports
(which can/will grow over time as future tests are added).