Improve non-crypto PRNG stuff
We had 32 and 64 -bit PRNGs defined.  The two usecases in the
codebase really only need 32-bit values.  One used the 32-bit PRNG
modulo a dynamic upper bound that was known-small and ignored the
tiny bias, and the other used the more-expensive 64-bit PRNG
because its dynamic upper bound is allowed to be relatively large
compared to 2^32.  In both cases bias was present but minimized,
and in both cases one division (due to modulo) was required on
every call.

This removes the 64-bit PRNG and provides a generic interface to
the 32-bit one that gives truly unbiased results regardless of the
magnitude or value of the bound and will on average perform better
in all senses for both cases.  It sometimes uses a single mod
division, and sometimes runs the (faster) core PRNG more than once
for a single result, but both are statistically unlikely.

The methods used here come from:

https://lemire.me/blog/2016/06/27/a-fast-alternative-to-the-modulo-reduction/
https://lemire.me/blog/2016/06/30/fast-random-shuffling/