Bugfix for $INCLUDE without origin argument
The testsuite was only exercising $INCLUDE with an explicit
origin-switching argument, and there was a bug when processing
$INCLUDE without the argument iff some previous record in the same
file was a type that used a domainname on the right hand side of
the record (e.g. an MX record).  The implicit origin of $INCLUDE
comes from the same internal data field and it wasn't reset before
processing, causing e.g. this fragment:

    foo MX 0 asdf
    $INCLUDE bar

to be effectively interpreted like this:

    foo MX 0 asdf
    $INCLUDE bar asdf