Refactor UDP EDNS max setting at dnspacket layer
Functional no-op.  Kills one pointless branch in the path of UDP
requests, and makes it explicit internally that the dnspacket
layer doesn't really care whether a given request came in over
IPv6 or not.  That information was only used for setting the max
size of UDP EDNS responses, which can be done at thread startup
rather than per-request.  This is necessary for correctness if
we're to support mixed IPv4 and IPv6 source addresses for certain
special TCP sockets anyways.