Move excess warning flags out of default build
The excessive warning flags are nice for me when doing
development, but it's pointless spam on other compilers for
production builds and wastes a lot of configure output log space
if nothing else.

As of now, the excessive warnings are only turned on if
--enable-developer or if the new flags --enable-extrawarn is set
on a non-developer build.  The default warning set includes just
-Wall.