Revert "Warn in docs and code about 2019 GeoLite2 changes"
This reverts commit 2415dab405693fe99c6fe2e14faeecd128692e93.

Bug #173

Apparently MaxMind retracted their intended removal of GeoLite2
coordinate data back in April 2018:

https://blog.maxmind.com/2018/04/11/we-are-listening-latitude-and-longitude-will-remain-in-geolite2-databases/

I never noticed or heard about it until it bit someone and they
filed a gdnsd bug.  We'll want to backport this and create a 2.4.1
release for users that can't upgrade to 3.x or aren't ready to
yet.  This may take a few days, since it's worth trawling through
some of the other 3.x work to see if there are any other trivial
bugfixes to backport while we're at it.