Fix EADDRNOTAVAIL issue under FreeBSD when using IPs bound to loopbacks
Without this patch, we pass the cmsg structure from recvmsg() directly
to sendmsg(), which results in an incorrect ifindex being set on packets
destined to loopback interfaces.

Reset ifindex to 0 if we are not link-local to allow the kernel to
choose the correct interface index.

(cherry-picked from 2.4.x branch, with some forward-porting cleanup,
and applied to the mainloop_mmsg() case as well, since that's newly
available on FreeBSD in 3.x --bblack).