Refactoring to meet most Sonar standards
I've started experimenting with using SonarScan via SonarCloud to
do some static analysis, coverage, and general QA work.  Like
Coverity, they're a commercial system that's free for actual open
source projects.  Unlike Coverity, they seem to not be having
constant downtime for the past month or two :)

Thanks to my colleague Guillaume Lederrey at Wikimedia for
suggesting the tool and doing an initial run last week in order to
nerd-snipe me into this tool!  The analysis is visible at:
https://sonarcloud.io/dashboard?id=gdnsd_gdnsd

The initial run through Sonar of course turned up a ton of issues.
I've trawled through them all in detail and none represent actual,
real-world bugs or vulnerabilities.  In some cases I disagree with
their standards and/or rationale, but for the most part I've tried
to just comply in cases where it's simple enough to do so and
doesn't make the code notably worse.  In at least some cases the
re-thinking and re-factoring prompted by this process really did
improve the clarity and cleanliness of the code in key places.

This commit is a squash of all the easy no-op refactors I've done
to bring the code mostly into compliance, at least to the level of
it logging zero counts for Vulnerabilities and Bugs and most of
the easy "Code Smell" stuff fixed.  A couple of more complex
fixups are split to separate followup commits coming right after
this one.

As of this writing there are still 54 "Code Smells" being logged
for non-trivial cases where I haven't quite decided what to do yet
or the work will be a bit harder (for example, "cognitive
complexity" sorts of issues).

I've also customized the "Sonar way" C ruleset as the "gdnsd way"
to deactivate certain rules that I don't think make sense in the
scope of this project.  The rules I've removed are all "Code
Smell" -level rules.  Four of them are basically of the form
"don't use C feature X because it's scary", where X is some
feature this codebase usefully uses at present:

* Unions should not be used
* "restrict" should not be used
* Preprocessor operators '#' and '##' should not be used
* Functions should not be defined with a variable number of args

The only other rule I removed so far was the one about not
allowing commented-out code.  I get their rationale on that one,
but (a) I do like leaving some commented out manual-debugging bits
here and there and (b) most of our violations are actually example
code (as in comments that explain via example how to use an
interface), which I tend to view positively.

I also resolved one outstanding issue as "Wontfix", which was
their recommendation to replace readdir() with readdir_r().  They
push this on the general grounds of using reentrant functions
where possible, but in the particular case of readdir() the world
is actually moving in the other direction now, with Linux/glibc
and FreeBSD documentation now deprecating readdir_r() in favor of
readdir() for $reasons and expecting POSIX to eventually follow.

I'm sure there will be some more rules tweaking and further
cleanup refactors on some of the remaining Code Smell issues, but
it's all in pretty decent shape as a starting point for now.

Overall I'm pretty impressed with Sonar's stuff so far, and we'll
probably eventually integrate it via Travis (once I figure out all
the harder bits there given I may want to keep some other Travis
integrations going in the matrix as well!) and badge it up in the
README.md, etc.  I still hope Coverity Scan comes back to life
too, though, as I'd like to be routinely checking with both!