Refactor ltree search functions
These two extremely similar (but unforunately also deeply
different and perf-sensitive) functions have always had horrible
structural craziness and complexity.  This was cleaned up while
working on Sonar conformance, but due to the complexity/risk in
this area it seemed better to a separate commit from the other
more-trivial cleanups.