Rework %%ITEM%% search/replace
This tiny function is one of the few places where Sonar was
flagging a "real" issue, as the code really wasn't sanity-checking
any constraints against overrunnning the 1K stack buffer here.

The reason it wasn't a real-world issue is that the code is only
used in configuration parsing of admin-provided input and the 1K
buffer is more than enough for real inputs.

Still, it's better to just be correct, so I've replaced it with a
generic string substitution function in libgdnsd that's a bit
cleaner/saner, and ensured that the testsuite exercises it.