Reserved identifiers cleanup
This is a mostly mechanical cleanup of up some basics about not
infringing on reserved namespaces.  Most of it is getting rid of
identifiers with leading underscores, but we also had several
variables or struct members with the same names as standard
library functions such as "sin", "asin", and "exp", all of which
were cleaned up as well.  Along the way, stopped putting redundant
_union and _struct suffixes on various such definitions in
ltree.h.