Testsuite fixup for TCP testing on BSD
The TCP state tests were failing on FreeBSD because the test
wasn't accounting for the fact that the accf_dns filter prevents
half the connections (the illegitimate ones) from ever reaching
gdnsd and bumping stats counters.  With this patch, the test
should work correctly and reliably regardless of whether the
accf_dns and/or accf_data kernel modules are loaded.