FreeBSD docs/tests updates
Tested on FreeBSD 12 now (before was just 11.2), and updated docs
slightly to match.

Interestingly, while we had to disable the default compiler
hardening flags (specifically the PIE stuff) on FreeBSD 11.2 to
get a successful build, with FreeBSD 12's build tools we actually
have to use the PIE stuff to get a successful build or we run into
errors linking some shared libs.

Also, fixed another testsuite bug on FreeBSD: 004misc/019bad was
relying on accept filters to have the same results as Linux, but
since the introduction of 026tcp/053tcp_states we don't need this
arrangement to test edge cases anymore, so I've re-arranged the
execution sequence of 004misc/019bad so it works the same
regardless of accept filter modules being loaded.