Bugfix for zonefile rrset size checks
The return type here should have been size_t, but was instead
written as bool.  This function's purpose is to reject loading
zonefiles which contain rrsets that would generate oversized
(>~16KB) responses, and because of this bug it totally failed at
its job (not that anyone should configure such data anyways!).