Implement RFC 8490 DNS Stateful Operations (DSO)
As of this writing I don't know of any other public RFC 8490
implementations to look at for interop testing.  None of the other
major DNS implementations seem to have it in their latest master
branches, anyways.  Therefore this is a blind implementation from
the RFC, with only our own regression suite to test it against for
sanity, and if I've made any logical or interpretive errors in
reading the RFC, I likely made the same mistake in both places!

It's enabled by default, but can be disabled with the global
option "disable_tcp_dso" in case of any bad interop or other
misbehaviors.