Bugfix for states query with no states
If gdnsd was configured with zero possible dynamic states to
manage (meaning no "plugins" or DYNx records used at all), and the
control client requests state info (e.g. "gdnsdctl states"), the
daemon would deliberately exit with an error message (and a
backtrace if libunwind support is compiled in), either about an
erroneous zero-byte allocation or an assertion failure, depending
on whether it's a prod or devel build.  What was expected instead
was no daemon exit, and the output of an empty json block through
the control client.

As the control socket client can by design also stop the daemon on
request, there's no real escalation here (protecting the unix
socket via standard permissions is a given), but this is clearly
undesirable behavior!  The fix is to ensure that even if no states
are defined, the "max_states_len" variable is still calculated
(which holds the size of the json output, and is still non-zero
even in the case the output is something like "{ }").