gdnsdctl: remove redundant/misleading -i help str
It was decided at some point to stop calling gdnsdctl's -i option
"idempotent" since the term doesn't quite make sense in all the
cases it applies to, and instead called it "ignore dead daemon".
However, when the new help string was added and the docs were
updated, the old help string was never deleted from the source.