Fixup rrset size checks from refactoring
The bugfix in 84252e89 (shared by master and 3.0.x) had already
fixed this bug once, but the master-only code smell refactor in
b808e10f re-introduced half the bug, which is fixed here :)  I'm
kind of loath to introduce actual testsuite coverage for the
disallowed giant record sets though.  We'll need a more-general
solution in the future as all of this gets rewritten.