Refactor dnspacket lookup code a bit more
Previous refactors still looked fairly ugly to me.  This one
replaces the loop used for CNAME chain processing with recursion
through two levels of function call here and takes care of a few
other related cleanups.  Fundamentally the job this part of the
code does has a lot of irreducible state and complexity, some of
which will hopefully go away if we can hopefully kill CNAME
chaining in a future version.