ACME Challenges: default to zero TTL on the wire
Our config setting "acme_challenge_ttl" was used to set two
different notions of "TTL": the time until the ephemeral text
record expired from gdnsd's memory and ceased being a query
response, and also the TTL value emitted on the wire towards
recursors for such records.  It has a reasonable default of 600
and lower bound of 60 for the first meaning, and it was assumed
the latter meaning (DNS-level TTL) wasn't very important.

As it turns out, the DNS-level TTL of these records is important
in real world scenarios, and the only value that's going to
guarantee non-buggy cert issuance in all possible scenarios with
all possible legitimate ACME server implementations is 0.

The case where non-zero TTLs are problematic is when an attempt is
made to issue two certs serially in quick succession which share a
SAN/CN element (e.g. the common case of ECDSA + RSA certs for the
same CN and SAN set issued back-to-back), and the ACME provider
caches the received records for the indicated DNS TTL, and also
uses two different challenge values (one per cert) for the shared
SAN/CN element.

In such a case, the ACME provider will fetch, cache, and validate
the first challege for the shared name in the first cert, then
fail the validation for the second certificate immediately
aftwards because the second challenge isn't in the rrset the ACME
provider has already cached locally, even though it's present in
the authserver at that point.

Even a very small non-zero TTL would present a race condition that
may trigger this failure, unless the client side automation
(what's driving gdnsd's ACME records) were explicitly aware of
this issue and intentionally delayed the attempted issuance of the
second cert by enough time to let the ACME provider's cached rrset
expire after it was fetched.

Our fix is to create a new seperate config setting for the second
meaning of the original "acme_challenge_ttl" and call it
"acme_challenge_dns_ttl".  It defaults to zero and allows a range
of 0-3600.  With the new default, users shouldn't ever run into
the scenario above.  It's up to ACME providers to honor the zero
TTL on the records, especially if they're not reusing challenge
values for the same name in short time windows.

A few other minor refactorings of related ltree code (the macros
related to TTLs, etc) are slipped into this commit as well, but
all are trivial.

See also the LE ticket where all this was figured out:

https://community.letsencrypt.org/t/unable-to-issue-ecdsa-rsa-in-acmev2-staging-environment/90835