Reduce minimum min_ttl setting to zero
Long ago when min_ttl was introduced, it was given a lower bound
of 1, disallowing static TTLs having the value zero.  Since we now
know of at least one case where a true-zero TTL is useful in the
real world, we should allow it to be possible for static records
as well.