Enforce min_ttl properly for SOA TTLs
The SOA TTL was being clamped to be no larger than its own ncache
ttl, which is in turn clamped to min_ttl and max_ncache_ttl.
However, this leaves a hole in the clamping where an SOA TTL which
is less than its own ncache TTL is never clamped to the global
min_ttl, and this patch fixes that.