Add 3.0.1 stuff to master NEWS
The 3.0.1 NEWS entry should've been included here when 3.1.0 was
released, as 3.1.0 contained all of these fixes as well.  This
edits the NEWS history a bit to inject it where it originally
belonged so that the history of 3.2.0 and beyond make sense.
We'll need a similar NEWS-copying fixup in any future 3.1.x
branch, should a need arise for a 3.1.1 release.