No-op to exercise new Sonar rules
Sonar has published an update to their quality rules for C which
adds several new checks.  We're likely to fail some of them and
need to dig through and decide which are real issues and which are
just questions of differing standards (I can already tell that a
couple of them are likely to be in the latter category).  I can't
seem to trigger sonar to try a new ruleset with a temporary
feature branch, so I'm making this "add a blank line to a
document" commit to kick Sonar on the master branch and get it to
report all the new failures for investigation...