Clean up address struct comparisons
A new Sonar bug rule is flagged a few cases here about comparing
struct padding in socket address structures.  None of these cases
seem like they can cause real bugs, but that's only because the
rest of the code is careful about initialization of padding to
zero as well.  It's safer and better practice to do as Sonar
recommends in these cases and not actually (or seemingly) compare
padding bits.

In the three more-trivial cases, there was a comparison of 16 raw
bytes of IPv6 address (&uint8_t[16]) to "struct in6_addr", when
the correct comparison target should be to in6_addr's "s6_addr"
member, but either way it's all the same, just cleaner to do it
correctly.

In the non-trivial case, the socket import code that matches up
handed-off sockets with configured listen addresses during smooth
daemon takeover was doing a wholesale memcmp compare of our entire
"struct gdnsd_anysin_t" for equality, even though that itself can
contain junk padding bits.  This worked fine in practice only
because we always zero the allocations and thus the padding bits
(esp in the v4 case), but it's more correct to create a comparison
function that steps through the correct fields of the union for
proper comparison and doesn't rely on the zeroing of padding bits
elsewhere.

When making changes to comply with new rules (as opposed to making
changes to fix issues flagged by existing rules), it seems
difficult to pre-flight check using temporary branches at all, so
I'm not sure if this actually squishes the new bug indications or
not until we see how it flies on Sonar's master branch eval.