Fix for comparison logic in previous commit
The gdnsd_anysin_cmp() function added in the previous commit
e09864a8 has a logic bug; the initial if clause should have a
logical-and rather than logical-or.  Nothing else noticed this
because, again, the rest of the code initializes anysin_t's
correctly and therefore if either of the length or family match,
both always do.  You could argue for assertions here instead, but
this more-generic form seems more future-proof.