Bugfix: ACME challenges should not stall takeovers
The monitoring initialization code requires the default event loop
of the main process be empty in order to function correctly.
However, when the ACME challenge hand-off feature was added, it
began injecting the imported ACME expiration events into the loop
*before* the monitoring initialization.

This could result in a 10 minute (or whatever non-default value
you use for `acme_challenge_ttl`) stall during the startup of the
new daemon instance when doing a 'gdnsdctl replace', in the case
that live, un-expired ACME challenges were being handed off from
one daemon to the next.