Add skip_first feature for geoip
This adds a 'skip_first' feature flag that can be set per-resource
for plugin_geoip resources.  If this is set, the first datacenter
in the failover list for a given lookup will be skipped, allowing
the definition of a "second choice" resource using the same basic
map definition as the first choice.

Technically this can be used with the metafo plugin too, but it's
not very useful in that context.