Auto-raise the files ulimit if needed
This adds a basic calculation that estimates the file descriptor
needs for all of the network sockets, plus another 100 fds of
buffer to handle other miscellaneous needs, and then uses this
value to raise any soft ulimit (within the bounds of the hard
ulimit), and warns if any of that doesn't work out or the hard
limit isn't high enough.

This code is non-fatal when it doesn't work, and is meant to make
for better default behavior in scenarios such as when the
tcp_clients_per_thread and tcp_threads values might allow for up
to 4K connections, but the default ulimits in effect are say 1K
soft and 64K hard.  In this case it's much better if we raise our
soft limit than let it fail, although the ultimate answer in any
tricky case is for the administrator to raise it manually for the
daemon.

Bug #197