Simplify accept4 errno handling
In the past, we attempted to split out some of the "soft" errno
cases for accept4 to be silent about (only debug log output), in
part because it was known they can sometimes be spammy.

Now that we have both an acceptfail stat and ratelimited log
outputs for cases like these, let's keep things simpler and stop
drawing any distinction between the many possible errnos, other
than in very clear cases.

Bug #197