Special handling for accept4 E[MN]FILE
The particular case of EMFILE/ENFILE (machine or process fd limit
reached) deserves special handling because it has been observed
that the errno return does not clear the pending connection, and
if we just log this condition and return to the eventloop, we will
tightly cycle back to accept4 repeatedly getting the same errno
for the same pending connection over and over again.

With this change, we attempt to treat these errnos the same way we
treat reaching our internal soft limits on client connections per
thread: we shoot down the most-idle of the existing idle
connections to make room for another fd.

Note that it will not always be the case that there's an idle
connection to shoot down at all, because the process-level fd
limit could be consumed by connections in threads other than the
one entering this error path, which happens to have none.  In that
case we'll continue handling it as a generic accept4 error as
before.

Bug #197