Adjust default max_edns_response[_v6] to 1232
In light of all the MTU and fragmentation discussions and research
that have lead to https://dnsflagday.net/2020/ (which, to be
clear, this codebase is already compliant with in the basic
sense), a change of defaults and rationales makes sense here.

These values were previously set to 1410 and 1212, respectively,
but the new default is better supported by modern evidence and
practice.  We continue to advertise the value 1024 in our response
packets; the defaults being changed here are for the other
direction and effectively cap the value we receive from clients.