Bugfix for TCP connection management
This fixes a subtle bug in the TCP connection management code.

The counter thr->check_mode_conns can leak upwards in certain
corner cases.  When this happens, it causes the ev_idle watcher
that manages buffered pipelined requests to remain active when it
shouldn't, which causes the eventloop to no longer block on the
network when there's no work left to do, thus wasting CPU
resources.  It also prevents the daemon from exiting cleanly, as
the affected TCP i/o thread refuses to terminate when asked.

This commit adds the missing counter decrement to fix the primary
bug, and also ensures that in any such scenario, the ev_idle
watcher cannot prevent the eventloop from exiting when no other
watchers remain.

Unfortunately, it's rather difficult to write a reasonable and
reliable regression test for this bug, so one isn't included.