Minor bugfix for static ACME challenge TTLs
If the configured acme_challenge_dns_ttl is less than the
configured min_ttl (which is the case with the defaults of 0 and
5, respectively), and a static acme challenge TXT record exists in
a zonefile which has the same TTL as the acme_challenge_dns_ttl,
the value would be clamped to min_ttl, which was not intended to
happen.  This was due to a simple logic bug in an if/else clause,
which is now fixed.

(cherry picked from commit 5a85dc1df9a0f56ce8ad293016114d5076bfc44b)