Emit "glue" for zone-level, non-delegation NS RRs
All of the 3.x releases to date have avoided emitting any additional
section address records when responding to a qtype=NS query at the
root of a zone.  This was done in the name of general response
minimization, and on the hypothesis that all such cases were
superfluous in practice in various ways (since the record could
and should otherwise be queried and/or cached separately), or
supported bad zone configuration habits that should be probably
avoided for most zones.

However, that seems in retrospect to have been an overzealous
optimization!  There are some such cases where the glue-like
outputs are useful, and there are others where it's simply not the
job of the authserver to decide whether a configuration is in good
taste or not.  In any case, sending them here doesn't get in the
flow of common user-initiated queries (as opposed to caches
updating metadata), and for the in-zone (delegated or not) cases
the cache would eventually want to query or refresh this data
anyways, and so sending it along in the additional section isn't
useless.

The most critical oversight was that gdnsd could no longer be
configured to serve root hints for qname='.',qtype='NS' when the
NS hostnames were in delegated subzones, as is the case for how
root hints are refreshed from the Internet's root servers, on the
off chance anyone ever tried to configure gdnsd in that role.

This commit returns us to emitting what looks like glue in all
such cases (although it's still technically not really necessary
glue in most of them).  The code still retains its disdain for and
deprecation warnings about truly out-of-zone glue for any NS
records, delegation or otherwise, but continues to emit them in
3.x for backwards compatibility.

Additionally, I've snuck into this commit a related change to a
source code comment in dnspacket.c, which incorrectly described
how the name compression code interacted with the root zone.